% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{bkde}
\alias{bkde}
\title{Estima a distribuicão bivariada de uma amostra por KDE}
\usage{
bkde(sp, h = c("NS", "STE"), kernel = dmvnorm, ngrid = 100)
}
\description{
Estima a distribuicão bivariada de uma amostra por KDE
}
