% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{contour.bkde}
\alias{contour.bkde}
\title{Plota um gráfico de contorno de BKDE}
\usage{
\method{contour}{bkde}(bkde, xlab = "x", ylab = "y", quantiles = c(0,
  0.25, 0.5), drawlabels = FALSE, ...)
}
\arguments{
\item{bkde}{objeto da classe bkde}

\item{quantiles}{defini os niveis do gráfico de contorno}
}
\description{
Plota um gráfico de contorno de BKDE
}
