% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{kde}
\alias{kde}
\title{Estima a distribuicão univariada de uma amostra por KDE}
\usage{
kde(sp, h = c("NS", "STE"), kernel = dnorm, ngrid = 10000)
}
\arguments{
\item{sp}{vetor numérico de qual o kde irá estimar a densidade}

\item{h}{método utilizado para o cálculo de h. A funcao possui duas opcões
, "STE" (solve the equation) e "NS" (normal scale). default = STE}

\item{kernel}{kernel utilizado na estimacão. default = dnorm}

\item{ngrid}{numero de pontos na malha. default = 10000}
}
\description{
Estima a distribuicão univariada de uma amostra por KDE
}
