% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_density.R
\name{kde_2d}
\alias{kde_2d}
\alias{print.kde_2d}
\alias{plot.kde_2d}
\title{Two Dimensional Kernel Density Estimate}
\usage{
kde_2d(
  x,
  y,
  N = 128,
  bw = c(x = bw.nrd0(x), y = bw.nrd0(y)),
  lower_x = limits$x$lower,
  upper_x = limits$x$upper,
  lower_y = limits$y$lower,
  upper_y = limits$y$upper,
  algorithm1 = length(x) < 2 * Nxy,
  k = 1
)

\method{print}{kde_2d}(x, ...)

\method{plot}{kde_2d}(
  x,
  ...,
  xlab = bquote(italic(.(x$data_name[["x"]]))),
  ylab = bquote(italic(.(x$data_name[["y"]]))),
  col = hcl.colors(25, rev = TRUE),
  restore_graphics_state = TRUE
)
}
\arguments{
\item{x}{Numeric vector or, in generics, a kde object}

\item{y}{Numeric vector of the response Variables}

\item{N}{Integer: number of equally spaced points for the kde.  May be one or two for x and y respectively}

\item{bw}{Numeric: bandwidth(s) for x and y kernels respectively.  A single value will be duplicated.}

\item{lower_x, upper_x, lower_y, upper_y}{Numeric: kde range limits}

\item{algorithm1}{Logical: should the smaller sample algorithm be used?}

\item{k}{Ineger: number of bandwidths by which to increase the range for the default limits.}

\item{...}{Extra arguments passed on to methods.}

\item{xlab, ylab, col}{base graphics parameters}

\item{restore_graphics_state}{logical: should the prior graphics
state be restored?}
}
\value{
An object of class "kde_2d"
}
\description{
A pure R implementation of two direct algorithms.
}
\examples{
set.seed(1234)
.time <- rnorm(10000, 10)
.money <- runif(.time) + .time/10
kxy <- kde_2d(.time, .money, algorithm1 = TRUE)
kxy
plot(kxy, restore_graphics_state = FALSE)
contour(kxy, add = TRUE)
rm(.time, .money)
}
