% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{plot.bkde}
\alias{plot.bkde}
\title{Plota os dados de BKDE}
\usage{
\method{plot}{bkde}(bkde, xlab = "x", ylab = "y", ...)
}
\arguments{
\item{bkde}{objeto da classe bkde}

\item{...}{você pode utilizar outros argumentos da funcão plot.default
xy <- rbnormmix(100, list(c(-5, -5), c(5, 5)))

bmix_bkde <- bkde(xy, h='STE')

plot(bmix_bkde,pch=16)
contour(bmix_bkde,add=TRUE)}
}
\description{
Plota os dados de BKDE
}
\examples{
 xy <- rbnormmix(100, list(c(-5, -5), c(5, 5)))

bmix_bkde <- bkde(xy, h='STE')

plot(bmix_bkde,pch=16)
contour(bmix_bkde,add=TRUE)
}
