% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{rnormmix}
\alias{rnormmix}
\title{Amostra uma mistura de normais univariada}
\usage{
rnormmix(n, mu, sigma = rep(1, length(mu)), prob = rep(1/length(mu),
  length(mu)))
}
\description{
Amostra uma mistura de normais univariada
}
