% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{bkde}
\alias{bkde}
\title{Estima a distribuicao bivariada de uma amostra por KDE}
\usage{
bkde(sp, h = c("NS", "STE"), kernel = dmvnorm, ngrid = 100)
}
\arguments{
\item{sp}{vetor numerico de qual o bkde ira estimar a densidade}

\item{h}{metodo utilizado para o calculo de h. A funcao possui duas opcoes
, "STE" (solve the equation) e "NS" (normal scale). default = STE}

\item{kernel}{kernel utilizado na estimacao. default = dmvnorm}

\item{ngrid}{numero de pontos na malha. default = 100}
}
\description{
Estima a distribuicao bivariada de uma amostra por KDE
}
\examples{
xy <- rbnormmix(100, list(c(-5, -5), c(5, 5)))

bmix_bkde <- bkde(xy, h='STE')

plot(bmix_bkde,pch=16)
contour(bmix_bkde,add=TRUE)
}
