% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{contour.bkde}
\alias{contour.bkde}
\title{Plota um grafico de contorno de BKDE}
\usage{
\method{contour}{bkde}(bkde, xlab = "x", ylab = "y", quantiles = c(0,
  0.25, 0.5), drawlabels = FALSE, ...)
}
\arguments{
\item{bkde}{objeto da classe bkde}

\item{quantiles}{defini os niveis do grafico de contorno}
}
\description{
Plota um grafico de contorno de BKDE
}
\examples{
xy <- rbnormmix(100, list(c(-5, -5), c(5, 5)))

bmix_bkde <- bkde(xy, h='STE')

plot(bmix_bkde,pch=16)
contour(bmix_bkde,add=TRUE)
}
