% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{dnormmix}
\alias{dnormmix}
\title{Densidade da mistura de normais univariada}
\usage{
dnormmix(x, mu, sigma = rep(1, length(mu)), prob = rep(1/length(mu),
  length(mu)))
}
\arguments{
\item{mu}{vetor das médias}

\item{sigma}{vetor da variâncias}

\item{n}{número de onservacões}

\item{prop}{vetor das probabilidades}
}
\description{
Densidade da mistura de normais univariada
}
