% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{dnormmix}
\alias{dnormmix}
\title{Densidade da mistura de normais univariada}
\usage{
dnormmix(x, mu, sigma = rep(1, length(mu)), prob = rep(1/length(mu),
  length(mu)))
}
\arguments{
\item{x}{vetor de quantis}

\item{mu}{vetor das médias}

\item{sigma}{vetor da variâncias}

\item{prop}{vetor das probabilidades}
}
\description{
Densidade da mistura de normais univariada
}
\examples{
sim <- rnormmix(105, mu = c(-1.75, 0.6, 3.0),
sigma = c(0.75, 0.5, 0.75), prob = c(3/6, 1/6, 2/6))

dsim <- normmix(sort(sim1), mu = c(-1.75, 0.6, 3.0),
sigma = c(0.75, 0.5, 0.75), prob = c(3/6, 1/6, 2/6))

plot(sort(sim),dsim, type = 'l')
}
