% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_density.R
\name{kde}
\alias{kde}
\alias{plot.kde}
\alias{print.kde}
\title{Kernel Density Estimate}
\usage{
kde(
  y,
  bw = bw.nrd0(y),
  N = 512,
  lower = ry[1] - 3 * bw,
  upper = ry[2] + 3 * bw
)

\method{plot}{kde}(
  x,
  ...,
  col = "steel blue",
  xlab = bquote(x == italic(.(x$data_name))),
  ylab = expression(kde(italic(x))),
  restore_graphics_state = TRUE
)

\method{print}{kde}(x, ...)
}
\arguments{
\item{y}{Numeric vector of observed values}

\item{bw}{Numeric value: bandwidth, standard deviation of the kernel density}

\item{N}{integer: number of equally spaced points for the kde}

\item{lower}{Numeric value: lower range limit for the kde}

\item{upper}{Numeric value: upper range limit for the kde}

\item{...}{Extra arguments passed on to methods.}

\item{xlab, ylab, col}{base graphics parameters}

\item{restore_graphics_state}{logical: should the prior graphics
state be restored?}

\item{an}{obeject of class "kde"}
}
\value{
a list object with results of the kde estimation.
}
\description{
A pure R implementation of a one-dimensional KDE with a
gaussian kernel.  Mainly of programming interest.
}
\examples{
set.seed(1234)
.x <- rchisq(1000, 2)
kx <- kde(.x, lower = 0)
kx
plot(kx, restore_graphics_state = FALSE, ylim = c(0,0.5))
curve(dchisq(x, 2), add = TRUE, col = "grey", n = 500)
rug(.x, col = 2)
rm(.x)
}
