% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_density.R
\name{kde}
\alias{kde}
\alias{plot.kde}
\alias{print.kde}
\alias{as.data.frame.kde}
\title{Kernel Density Estimate}
\usage{
kde(
  y,
  bw = bw.nrd0(y),
  N = 512,
  lower = ry[1] - 3 * bw,
  upper = ry[2] + 3 * bw
)

\method{plot}{kde}(
  x,
  ...,
  col = "steel blue",
  xlab = bquote(x == italic(.(x$data_name))),
  ylab = expression(kde(italic(x)))
)

\method{print}{kde}(x, ...)

\method{as.data.frame}{kde}(x, row.names, optional, ..., responseName = "density")
}
\arguments{
\item{y}{Numeric vector of observed values}

\item{bw}{Numeric value: bandwidth, standard deviation of
the kernel density}

\item{N}{integer: number of equally spaced points for the kde}

\item{lower}{Numeric value: lower range limit for the kde}

\item{upper}{Numeric value: upper range limit for the kde}

\item{x}{an obeject of class "kde"}

\item{...}{Extra arguments passed on to methods.}

\item{xlab, ylab, col}{base graphics parameters}

\item{row.names, optional}{As for the as.data.frame generic}

\item{responseName}{character string: name for the response
variable}
}
\value{
a list object with results of the kde estimation.
}
\description{
A pure R implementation of a one-dimensional KDE with a
gaussian kernel.  Mainly of programming interest.
}
\examples{
set.seed(1234)
x_value <- rchisq(10000, 2)
kx <- kde(x_value, lower = 0)
kx
if("package:ggplot2" \%in\% search()) {
  ggplot(as.data.frame(kx)) +
    aes(x = x_value, y = density) +
    geom_line(colour = "steel blue") +
    geom_function(fun = function(x) dchisq(x, 2),
                  colour = "firebrick3") +
    theme_bw() +
    xlab(expression(x == italic(x_value))) +
    ylab(expression(kde(italic(x))))
} else {
  oldPar <- par(mar = c(3, 3, 1, 1) + 0.1, las = 1, mgp = c(1.5, 0.5, 0),
                cex.axis = 0.7, cex.lab = 0.8, tck = -0.015)
  plot(kx, ylim = c(0,0.5))
  curve(dchisq(x, 2), add = TRUE, col = 2, n = 1000)
  rug(x_value, col = "dark green")
  box()
  par(oldPar)
}
rm(oldPar)
}
