% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{kde}
\alias{kde}
\title{Estima a distribuicão univariada de uma amostra por KDE}
\usage{
kde(sp, h = c("NS", "STE"), kernel = dnorm, ngrid = 10000)
}
\description{
Estima a distribuicão univariada de uma amostra por KDE
}
