% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{kde}
\alias{kde}
\title{Estima a distribuicao univariada de uma amostra por KDE}
\usage{
kde(sp, h = c("NS", "STE"), kernel = dnorm, ngrid = 10000)
}
\arguments{
\item{sp}{vetor numerico de qual o kde ira estimar a densidade}

\item{h}{metodo utilizado para o calculo de h. A funcao possui duas opcoes
, "STE" (solve the equation) e "NS" (normal scale). default = NS}

\item{kernel}{kernel utilizado na estimacao. default = dnorm}

\item{ngrid}{numero de pontos na malha. default = 10000}
}
\description{
Estima a distribuicao univariada de uma amostra por KDE
}
\examples{
x <- rnormmix(100, c(-5, 5))
xy <- kde(x, h='STE')

plot(xy)
}
