% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{p_fun}
\alias{p_fun}
\alias{p_fun.kde}
\alias{p_fun.numeric}
\alias{d_fun}
\alias{d_fun.kde}
\alias{d_fun.numeric}
\alias{q_fun}
\alias{q_fun.kde}
\alias{q_fun.numeric}
\alias{r_fun}
\alias{r_fun.kde}
\alias{r_fun.numeric}
\title{Kernel Distribution Functions}
\usage{
p_fun(x, ...)

\method{p_fun}{kde}(x, ...)

\method{p_fun}{numeric}(x, ...)

d_fun(x, ...)

\method{d_fun}{kde}(x, ...)

\method{d_fun}{numeric}(x, ...)

q_fun(x, ...)

\method{q_fun}{kde}(x, ...)

\method{q_fun}{numeric}(x, ...)

r_fun(x, ...)

\method{r_fun}{kde}(x, ...)

\method{r_fun}{numeric}(x, ...)
}
\arguments{
\item{x}{Either a kde object or a numeric vector from which a kde
object can be generated.}

\item{...}{Additional arguments forwarded to the \code{kde} function.}
}
\value{
An R function for probabilities, densities, quantiles or
         simulated samples, as indicated by the prefix.
}
\description{
A set of functions to generate p-, d-, q- or r-prefix functions,
as are available for standard probability distributions, but for
distributions defined by a kernel density estimate.
}
\examples{
set.seed(1234)
x <- rnorm(10000)
kx <- kde(x)
pkx <- p_fun(kx)
dkx <- d_fun(kx)
qkx <- q_fun(kx)
rkx <- r_fun(kx)
rbind(quantile(x, 2:8/10), qkx(2:8/10))
pkx(qkx(1:9/10))
curve(dkx, xlim = qkx(0:1), col = 2)
curve(dnorm, add = TRUE)
grid()
rx <- rkx(x)
qqplot(x, rx, pch = ".")
abline(0, 1, col=2)
grid()
rm(x, rx)
}
