% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{persp.bkde}
\alias{persp.bkde}
\title{Plota a densidade estimada por BKDE}
\usage{
\method{persp}{bkde}(bkde, xlab = "x", ylab = "y", zlab = "density",
  ticktype = "detailed", ...)
}
\arguments{
\item{bkde}{objeto da classe bkde}
}
\description{
Plota a densidade estimada por BKDE
}
\examples{
xy <- rbnormmix(100, list(c(-5, -5), c(5, 5)))

bmix_bkde <- bkde(xy, h='STE')

persp(bmix_bkde,theta=-120,phi=30)
}
