% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{rnormmix}
\alias{rnormmix}
\title{Amostra uma mistura de normais univariada}
\usage{
rnormmix(n, mu, sigma = rep(1, length(mu)), prob = rep(1/length(mu),
  length(mu)))
}
\arguments{
\item{n}{número de observacões}

\item{mu}{vetor das médias}

\item{sigma}{vetor da variâncias}

\item{prop}{vetor das probabilidades}
}
\description{
Amostra uma mistura de normais univariada
}
